atualização

+++ ordem/application/controllers/Usuarios.php
@@ -1,11 +1,10 @@
 <?php
+
 defined('BASEPATH') or exit('No direct script access allowed');
 
-class Usuarios extends CI_Controller
-{
+class Usuarios extends CI_Controller {
 
-    public function __construct()
-    {
+    public function __construct() {
         parent::__construct();
 
         if (!$this->ion_auth->logged_in()) {
@@ -14,23 +13,23 @@ class Usuarios extends CI_Controller
         }
     }
 
-    public function index()
-    {
+    public function index() {
+
+        if (!$this->ion_auth->is_admin()) {
+            $this->session->set_flashdata('info', 'Você não possui permissão para acessar o menu Usuários');
+            redirect('/');
+        }
 
         $data = array(
-
             'titulo' => 'Usuários cadastrados',
-
             'styles' => array(
                 'vendor/datatables/dataTables.bootstrap4.min.css'
             ),
-
             'scripts' => array(
                 'vendor/datatables/jquery.dataTables.min.js',
                 'vendor/datatables/dataTables.bootstrap4.min.js',
                 'vendor/datatables/app.js'
             ),
-
             'usuarios' => $this->ion_auth->users()->result(),
         );
 
@@ -41,8 +40,12 @@ class Usuarios extends CI_Controller
         $this->load->view('layout/footer');
     }
 
-    public function add()
-    {
+    public function add() {
+
+        if (!$this->ion_auth->is_admin()) {
+            $this->session->set_flashdata('info', 'Você não possui permissão para acessar o menu Usuários');
+            redirect('/');
+        }
 
         $this->form_validation->set_rules('first_name', '', 'trim|required');
         $this->form_validation->set_rules('last_name', '', 'trim|required');
@@ -90,32 +93,17 @@ class Usuarios extends CI_Controller
         }
     }
 
-    public function edit($usuario_id = NULL)
-    {
+    public function edit($usuario_id = NULL) {
+
+        if ($this->session->userdata('user_id') != $usuario_id && !$this->ion_auth->is_admin()) {
+            $this->session->set_flashdata('error', 'Você não pode editar um usuário diferente do seu');
+            redirect('/');
+        }
 
         if (!$usuario_id || !$this->ion_auth->user($usuario_id)->row()) {
-
             $this->session->set_flashdata('error', 'Usuário não encontrado');
             redirect('usuarios');
         } else {
-
-            /*
-                [first_name] => Admin
-                [last_name] => istrator
-                [email] => admin@admin.com
-                [username] => administrator@gmail.com
-                [active] => 1
-                [perfil_usuario] => 1
-                [password] => asd
-                [confirm_password] => 
-                [usuario_id] => 1
-            */
-
-            /*
-            echo '<pre>';
-            print_r($this->input->post());
-            echo exit(); 
-            */
 
             $this->form_validation->set_rules('first_name', '', 'trim|required');
             $this->form_validation->set_rules('last_name', '', 'trim|required');
@@ -126,18 +114,25 @@ class Usuarios extends CI_Controller
 
             if ($this->form_validation->run()) {
 
-                $data = elements(
+//                echo '<pre>';
+//                print_r($this->input->post());
+//                exit();
 
-                    array(
-                        'first_name',
-                        'last_name',
-                        'email',
-                        'username',
-                        'active',
-                        'password'
-                    ),
-                    $this->input->post()
+                $data = elements(
+                        array(
+                            'first_name',
+                            'last_name',
+                            'email',
+                            'username',
+                            'active',
+                            'password'
+                        ),
+                        $this->input->post()
                 );
+
+                if (!$this->ion_auth->is_admin()) {
+                    unset($data['active']);
+                }
 
                 $data = $this->security->xss_clean($data);
 
@@ -145,7 +140,6 @@ class Usuarios extends CI_Controller
 
                 //Verifica se foi passado o password
                 if (!$password) {
-
                     unset($data['password']);
                 }
 
@@ -155,19 +149,23 @@ class Usuarios extends CI_Controller
 
                     $perfil_usuario_post = $this->input->post('perfil_usuario');
 
-                    /* Se for diferente atualiza no banco */
-                    if ($perfil_usuario_post != $perfil_usuario_db->id) {
-
-                        $this->ion_auth->remove_from_group($perfil_usuario_db->id, $usuario_id);
-                        $this->ion_auth->add_to_group($perfil_usuario_post, $usuario_id);
+                    if ($this->ion_auth->is_admin()) {
+                        /* Se for diferente atualiza no banco */
+                        if ($perfil_usuario_post != $perfil_usuario_db->id) {
+                            $this->ion_auth->remove_from_group($perfil_usuario_db->id, $usuario_id);
+                            $this->ion_auth->add_to_group($perfil_usuario_post, $usuario_id);
+                        }
                     }
-
                     $this->session->set_flashdata('sucesso', 'Dados salvos com sucesso');
                 } else {
                     $this->session->set_flashdata('error', 'Erro ao salvar os dados');
                 }
 
-                redirect('usuarios');
+                if (!$this->ion_auth->is_admin()) {
+                    redirect('usuarios');
+                } else {
+                    redirect('/');
+                }
             } else {
 
                 $data = array(
@@ -183,30 +181,30 @@ class Usuarios extends CI_Controller
         }
     }
 
-    public function del($usuario_id = NULL)
-    {
+    public function del($usuario_id = NULL) {
+
+        if (!$this->ion_auth->is_admin()) {
+            $this->session->set_flashdata('info', 'Você não possui permissão para acessar o menu Usuários');
+            redirect('/');
+        }
 
         if (!$usuario_id || !$this->ion_auth->user($usuario_id)->row()) {
 
             $this->session->set_flashdata('error', 'Usuário não encontrado');
             redirect('usuarios');
-
-        } else if($this->ion_auth->is_admin($usuario_id)){
+        } else if ($this->ion_auth->is_admin($usuario_id)) {
 
             $this->session->set_flashdata('error', 'Administrador não pode ser excluído');
             redirect('usuarios');
-
-        } else if($this->ion_auth->delete_user($usuario_id)){
+        } else if ($this->ion_auth->delete_user($usuario_id)) {
 
             $this->session->set_flashdata('sucesso', 'Usuário excluído com sucesso');
             redirect('usuarios');
-            
         }
     }
 
     //Verifica se e-mail já existe
-    public function email_check($email)
-    {
+    public function email_check($email) {
 
         $usuario_id = $this->input->post('usuario_id');
 
@@ -222,8 +220,7 @@ class Usuarios extends CI_Controller
     }
 
     //Verifica se username já existe
-    public function username_check($username)
-    {
+    public function username_check($username) {
 
         $usuario_id = $this->input->post('usuario_id');
 
@@ -237,4 +234,5 @@ class Usuarios extends CI_Controller
             return TRUE;
         }
     }
+
 }
